memcached-tool dump command fixed for expired/evicted/deleted keys
Fixed the memcached-tool 'dump' command to handle the case when
dumping keys that have expired, evicted or been deleted in the midst
of the big iteration.  Previously, memcached-tool would just hang in
these cases.  The memcached-tool uses the results from 'stats
cachedump' as the keys to iterate through, and will do a 'get' on each
key.  The 'stats cachedump' implementation, however, does not check
for expiry, which is debatably the correct thing to do.