Return all errors for binary commands except not_found for getk.
The theory is that we want quiet commands to be quiet in *normal*
cases.  A set that succeeds does not need to return anything.

A quiet get has two normal cases:

1) Got a value (must return it!)
2) not_found (quiet here)

All other errors are potentially interesting, so they will also be
allowed through.

Client authors can use this information to create bath commands of
various types and not lose any useful status.  A multi-set can be as
efficient as a multi-get and still know which sets worked and which
failed (even in the case of CAS).