Get build working under ICC.
ICC pretends to be GCC as far as autoconf is concerned, but is
incompatible in a few ways.

ICC C99 mode fails to define u_char, so I made a small change to
modify that required c99 to work with ICC's C99 mode off.  Nobody
wants to put effort into working around compilers that don't speak C99
in the long-term, but a one line change has already revealed quite a
few potential bugs.