Remove bad memset call in process_bin_stats.
(Victor found the bug and submitted a bunch of text, I just committed
it -- Dustin)

Unclear why this exists in the first place (history doesn't show
anything significant), but at some point after it was written, it
became wrong.

As this is a local variable and is not read after this memset and the
memset is wrong, I'm taking Victor's path of just removing it.

http://code.google.com/p/memcached/issues/detail?id=19

Signed-off-by: Dustin Sallings <dustin@spy.net>