Isolate the first server in startup test so it dies sooner (bug 20).
Apparently on some version of Linux I don't have, this causes the test
to hang for a while as gcov attempts to compute and write out the
results from two coverage runs concurrently.

It's best to clean up things as quickly as possible in the general
case, so this should be done regardless.