Don't avoid binding UDP on IPv6.
There was apparently some issue on OS X 10.5 that led to the
introduction of this limitation, but it's not clear what the issue
was, and all tests pass on OS X 10.5 for me.

Assuming we find out what the cause of this issue is, we'll add tests
and sure they work fine on all platforms with no (or minimal) loss of
functionality.