dumb hack to self-repair stuck slabs
since 1.2.6, most of the refcount leaks have been quashed.
I still get them in production, extremely rarely.
It's possibly we'll have refcount leaks on and off even in the future.

This hack acknowledges this and exists since we want to guarantee, as much as
possible, that memcached is a stable service. Having to monitor for and
restart the service on account of "rare bugs" isn't acceptable.