fix resumption of accept loop under multithread
critical fix:
under multithreaded mode, in version 1.2.7, memcached would not resume
accepting connections after hitting the max connections limit.

this is due to a pre-existing bug in the accept_new_conns code, which
was hit when the "fix" was rolled in to ensure listening threads never
did real work.  Previously, at least one closing connection would
(randomly) be on the accepting thread, so the test for (are we the
listening thread?) would still work.