Fix for issue 72: 1.4.0 binary multi-get crash issue
Repack the input buffer if the current command and key doesn't fit in the
input buffer (may occur if we read multiple commands in a pipeline)

(Dustin's note):

The original test was connecting to "localhost" which was resolving to
::1 on a few machines.  With the ephemeral port binding, IPv4 and IPv6
were getting different ports, and the code was explicitly looking for
the IPv4 port, but implicitly connecting to ::1.  Now it explicitly
connects to 127.0.0.1 over IPv4.