Allow memcached to bind to an ephemeral port
To ease testing you can pass -p -1 to memcached and it will create and bind
to an ephemeral port. If you set the environment variable
MEMCACHED_PORT_FILENAME to a filename memcached will write the port numbers
to the file in the following format:
TCP INET6: 5555
TCP INET: 5555
UDP INET6: 6666
UDP INET: 6666

testapp can now just tell the server to start and don't try to locate
a random port for it to use, but instead read the file and connect on
the port.