Add support for hugetlbfs (for Linux).
From http://libhugetlbfs.ozlabs.org/ -

     libhugetlbfs is a library which provides easy access to huge
     pages of memory. It is a wrapper for the hugetlbfs file
     system.

If you are running memcached with a very large heap in Linux, this
change will make it available to you.  The hugetlbfs HOWTO provides
detailed information on how to configure your Linux system and provide
advice to applications (such as memcached) to make use of it.