Cleanup of number of threads declarations (issue 91)
     * Change setings.num_threads (-t option) meaning (Now it is a number
       of worker threads).
     * Fix bug in -t option checking.
     * Simple data struct for dispatcher (no thread-local stat and so on).
     * No special threads[0] for dispatcher thread info.
     * thread_local_stats_{reset|aggregate} does not cycle on unused
       dispatcher thread stat.
     * Simplify thread initialization and connection dispatching logic.

(notes from Dustin):

A list in a commit is typically a red flag, but this isn't really
listing a bunch of distinct things that were done, but a bunch of ways
things were made better by a simple refactoring.

I also added a test that verifies that it fails if you add "-t 0".
Before, it did not fail, but the whole server would crash if you
connected to it.  This test doesn't confirm the server crashed in that
case, but at least confirms the exact issue 91 case, which is that it
does the right thing when "-t 0" is specified.