Remove slab size special case for growth factor 2. bug56
Using any growth factor, the minimum slab size is always 104 bytes
(unless using -n option to use a specific value) except for 2. In this
case, it's 128 bytes and cannot be reduced using -n option (it can be
increased however).

From slabs.c:
    /* Factor of 2.0 means use the default memcached behavior */
    if (factor == 2.0 && size < 128)
        size = 128;

What I sugget is to simply remove this part. If the default behaviour
needs to be kept to 128 bytes per default when factor is 128, maybe
default value could for the -n option could be adapted to do so. But I
don't see a good reason for this specific difference between factor 2
and others.