Update stat documentation to reflect allocation strategy.
The documentation stated that memory allocation was always in chunks
of 1MB.  In practice, this gets less effecient as you get into larger
slabs.

The 1MB thing (which can be controlled by -I) is both the maximum size
and allocation unit so a single allocate event will attempt to get as
many items for a particular slab as it can as long as the allocation
itself doesn't exceed the maximum item size.