Allow semi-broken C99 compilers to compile memcached
Some C compilers understand the syntax we use in memcached, but do not
ship stdbool.h or stdint.h.  According to C99 inttypes.h contains the
formatting macros we use (PRInn) and is supposed to include stdint.h.

This patch tries to detect stdbool.h and inttypes.h and tries to
include them from config.h