Added auth_enabled_sasl setting stat and tests for the sasl stats.
Added basic sasl stats and tests for whether or not those stats
return the expected values.  The preferred method would be to reauth
but that is not expected to currently work because of how the underlying
sasl library works.  Instead, the test is dependent upon what has happened
earlier with other tests.

The name for the stat includes the suffix of _sasl because there may
some day be an option for other types of authentication.