Supress fwrite wur in mcstat.
mcstat pulls statistics from memcached and writes them to stdout.
Recent versions of glibc have a warn_unused_result attribute on fwrite
causing this utility to fail to compile.

While this change in glibc is a bit controversial[0], it's the state
of things on some systems today, so this change explicitly ignores the
result of writes to stdout.

[0]: http://sourceware.org/bugzilla/show_bug.cgi?id=11959