msghdr->msg_namelen should only be set for UDP
We are currently using sockaddr_storage to keep track of the address
(to support IPV6), but this caused sendmsg to fail on OpenBSD with
EINVAL. The socket is already in a connected state, so we don't
need to pass in the socket address.

Our test coverage is a bit low for UDP, so I leave the initalization
there for now. I'm pretty sure that we could just set this to zero
here, because we reset the member each time we receive a packet
in try_read_udp().