workaround low mtime resolution by reloading isasl db few times
If isasl.pw is changed several times during one second (which is
resolution of mtime in struct stat), then we'll have stale password
database.

To prevent that we keep reloading isasl password for 2 mtime check
cycles. If mtime is still same, we can be sure that we haven't missed
updates.

Change-Id: Ia299aba6296b86ec0e569cf1d13bc9fbede0303a