experimental maxconns_fast option
Also fixes -c option to allow reducing the maximum connection limit.

This gives a new option "-o maxconns_fast", which changes how memcached
handles hitting the maximum connection limit. By default, it disables the
accept listener and new connections will wait in the listen queue. With
maxconns_fast enabled, new connections over the limited have an error written
to them and are immediately closed by the listener thread.

This is currently experimental, as we aren't sure how clients will handle the
change. It may become the default in the future.