Remove the depth search from item_alloc
Code checked 50 items before checking up to 50 more items to expire one, if
none were expired. Given the shallow depth search (50) by any sizeable cache
(as low as 1000 items, even), I believe that whole optimization was pointless.

Flattening it to be a single test is shorter code and benches a bit faster as
it holds the lock for less time.

I may have made a mess of the logic, could be cleaned up a little.