Add missing LRU bumps to binary get and others
An audit turned up that the LRU bump to move just accessed items to the front
was missing from the binary get command. It was also missing from incr/decr
and the new touch commands.

If someone was using the binary protocol exclusively, memcached would be
acting as a FIFO for stored items.