disable issue 140's test.
the fix for issue 140 only helped in the case of you poking at memcached with
a handful of items (or this particular test). On real instances you could
easily exhaust the 50 item search and still come up with a crap item.

It was removed because adding the proper locks back in that place is
difficult, and it makes "stats items" take longer in a gross lock anyway.