don't fetch yourself during memory allocation
It was possible to, while replacing an item, find your own item from the tail
and examine it. This reverse some logic and adds a not-completely-covering
check of ignoring an item if it has the same (complete) hash value.

This is the second half of the general fix for the infamous "issue 260" from
google code bugs.