Avoid OOM errors when locked items stuck in tail
If a client fetches a few thousand keys, then does not ever read the socket,
those keys will stay reflocked until the client disconnects or resumes. If
some of those items are unpopular they can drop to the tail, causing all
writes in the slab class to OOM.

This creates some relief by chucking the items back to the head.

Big thanks to Jay Grizzard and other folks at Box for helping narrow this
down.