incr/decr: the value should be part of extras according to doc
protocol-binary.xml states:
  Increment, Decrement
    MUST have extras.
    MUST have key.
    MUST NOT have value.

but memcached did not set extlen for incr and decr responses.  Most client
applications didn't appear to care, they just used the 8 bytes of the body
as the delta value and thus should not break when incr and decr start saying
the new value is in extras.