Fix potential corruption for incr/decr of 0b items
It wasn't checking that the data was nonzero in size. Internally it uses
strtoull, which ignores initial characters matched as spaces, then attempts to
parse the rest of the number.

An empty item has "\r\n" in the data section (min size 2), so if there's
parseable memory beyond the end of the data section in that chunk, possibly
from data previously occupying that space, it'll successfully delta it.

It won't corrupt memory, since it later tests for space to write the new value
into (it->nbytes), and will then allocate a new item for the storage.