compat mode.
Enabling the new LRU routine requires starting with `-o lru_maintainer`.

This makes almost all of the tests pass, except refhang.t. Will need new tests
for the LRU maintainer.

So far as I can tell it's still handling the refhang scenario, but in a more
natural way. Instead of flipping the items back to the top of the list, it's
unlinking them from the hash table and LRU. This completely removes them from
the problem, but it doesn't retry as many times to get them out of the way.

A system with many stuck items next to each other could do a handful of OOM's
before clearing the backlog, but it won't keep running into them. The test
appears flaky even in 1.4.22; running with -vv causes it to fail in a funny
way.