fix refhang test.
The new code is a lot more efficient as unblocking LRU's as it's able to
unlink refcounted items. However it's less aggressive in these cases. You'll
get one OOM per stuck item and then it'll be gone in most cases.

Removed the bottom half of the test since it's too flaky, and the above case
now looks for both OOM's and STORED's plus relevant counters.