spinlocks never seem to help in benchmarks
If a thread is allowed to go to sleep, it can be woken up early as soon as the
lock is freed. If we spinlock, the scheduler can't help us and threads will
randomly run out their timeslice until the thread actually holding the lock
finishes its work.

In my benchmarks killing the spinlock only makes things better.