relax timing glitch in the lru maintainer test
This test is requiring that the juggler thread runs at all before the stats
check happens. I've tried running this under an rPi1 and can't reproduce the
race, but for some reason solaris amd64 does. This is likely due to the usleep
not working as expected.

Unfortunately I don't have direct access to a solaris host, so this is the
best I can do for now. The juggler does eventually wake up so I'm unconcerned.