dump command learned to handle no-expire keys
memcache internals can use the timestamp generated by time - uptime from
stats to signal a no-expire key.  The dump command didn't account for
this.  With this change, the dump command now uses the stats command's
output to generate the no-expire timestamp.  Any exported timestamp that
matches this no-expire timestamp is replaced with a 0 instead, so that
when reimporting the data, the key is properly set as a no-expire key.