use one bipbuffer per watcher. remove iovec code.
after the previous commit's exercise, this greatly simplifies the writing
process. The buffers are large but can be tuned with performance testing.

Lines are now rendered into scratch space, and if event flags (eflags) match
for a watcher, copy the bytes into that watcher's specific buffer.

In most realistic cases, watchers will be watching different streams of
information. While this does require a minimum of one memcpy instead of
directly writing into the bipbuf as before, it removes all of the loops and
management of iovecs required for filtering events on the write-to-socket end.