add "-o modern" option for easy default updates
I've been recommending people use a long string of arguments as the canonical
way of starting memcached. This boils it down to a single argument that should
automatically pick up new stuff in the future as well.

These defaults will become the real defaults at some point, so I hope folks
are trying it out.