Fix undefined behavior in vlen check
The size check for vlen in process_get_command was relying on
undefined behavior (signed integer overflow) to check to see
if vlen was greater than 2^31-2. This made certain compilers
(such as clang 7.3.x) optimize out part of the conditional and
prevent memcache from erroring when the specified size was
exactly 2^31-1, leading to an incorrect error message and a core
dump. This patch does that overflow test in a not-undefined-behavior
way.

Fixes failure in issue_70.t test #4 on those platforms.