test hit ratio fix for LRU maintainer mode
under higher loads, the LRU maintainer might not get to peek at a tail before
an eviction is required, so active items at the tail can be pulled.

This attempts to address it in two ways: one, allow the LRU maintainer thread
to run even faster. Two, if you get a hit in cold, immediately move to warm.
This also stops pulling from HOT/WARM during an eviction cycle.

This is a long-missing part of the algorithm which has a more optimized route
of per-worker-thread queues of moving items to warm. That should re-enable
no-shuffling-during gets. In this test mode, shuffling only happens in the
relatively rare hits on cold. Hits on HOT/WARM are both shuffled by the LRU
maintainer thread.