fix race condition with stacks
hopefully I can find a less fragile way of dealing with IO stacks, but this
should fix it.

Under load tests it became obvious the callback could return _after_ the
worker thread was woken up and re-added an entire stack from scratch. because
I should know that.

leaving the debug code and extra asserts in during development.

the bench runs on my first test machine for a couple minutes now, and hasn't
crashed.. but it does eventually hang. This is probably related to the other
race condition on reading the queue and listening on a condition.