extstore: not TTL aware, but use arbitrary buckets
since objects are held in memory, and the LRU crawler exists, expired objects
should get occasionally cleared. With the plan to add a function call to
inform a page of a deleted object, there's no reason to track aging internally
to extstore.

now uses arbitrary write buckets. will discuss that in the documentation.