memc: fix ordering issue in conn dispatch
very old race condition: push to queue, write to signal pipe. with extstore,
hammering a secondary item type into the queue could occasionally swap the
item with the type it represented.

now differentiate on the type, and the character signifies either to pull an
item or do something else entirely.