extstore: basic binary protocol support
under high memory pressure the protocol can desync due to a lack of error
handling code paths in binprot :( But; if that's rare enough this is testable.

like ascii, incr/decr/append/prepend and chunked items don't presently work.

change was tricky but much smaller than I thought it'd be.