extstore: automove algorithm test script
first pass at an extstore-friendly automove algorithm. With a static item_age
value, extra RAM may be unused during all or parts of the day, and the
existing balancer may not give enough memory to small slab classes.

This does a lot of page moving and can be further improved.
- leave some of the "buffer" as actual free space in each slab class. only
  move pages back to global if one gets way too much memory.
- use a PID controller to slew adjustments.

but it does work, so I'll probably convert this to C and revise the algo
later.