extstore: add ext_drop_unread option + live tune
was struggling to figure out how to automatically turn this on or off, but I
think it should be part of an outside process.

ie; a mechanism should be able to target a specific write rate, and one of
its tools for reducing the write rate should be flipping this on.

there's *still* a hole where you can't trigger a compaction attempt if
there's no fragmentation. I kind of want, if this feature is on, to attempt
a compaction on the oldest page while dropping unread items.