extstore: experimental per-class free chunk limit
external comands only for the moment. allows specifying per-slab-class how
many chunks to leave free before causing flushing to storage.

the external page mover algo in previous commits has a few issues:

- relies too heavily on page mover. lots of constant activity under load.
- adjusting the item age level on the fly is too laggy, and can easily
  over-frefree or under-free. IE; class 3 has TTL 90, but class 4 has TTL 60
and most of the pages in memory, it won't free much until it lowers to 60.

Thinking this would be something like; % of total chunks in slab class.
easiest to set as a percentage of total memory or by write rate periodically.

from there TTL can be balanced almost as in the original algorithm; keep a
small global page pool for small items to allocate memory from, and pull
pages from or balance between storage-capable classes to align TTL.