extstore: fix size tracking and adjust drop_unread
was early evicting from HOT/WARM LRU's for item headers because the
*original* item size was being tracked, then compared to the actual byte
totals for the class.

also adjusts drop_unread so it drops items which are currently in the COLD_LRU

this is expected to be used with very low compacat_under values; ie 2-5
depending on page count and write load. If you can't defrag-compact,
drop-compact.

but this is still subtly wrong, since drop_compact is now an option.