extstore: another page balancer tweak
wasn't aggressively enough reclaiming excess free memory if it was wobbling
around a specific number.

should drop it with some stop and go but will fix and should have a window
sized delay to absorb wobbles in free memory due to crawlers freeing things.

might be able to further tweak by preventing the reclaim if it's
*consistently* dropping for the whole window. a task for the future :P