extstore: fix crash while moving extstore chunks
ITEM_LINKED was still set on the objects being written to disk. If a page
being moved contains a chunk read from extstore currently being written back
to the client, it will mistake it for a chunk properly linked and attempt to
unlink if if the page has also been jammed in the mover.

So you need a cross section of a particular chunk being active the entire
time a page is jammed, then it tries to unlink it but the header is partially
garbage and it segfaults.

The page mover ignores all items which don't either have LINKED or SLABBED,
assuming they're in transition. So the fix is to simply remove the LINKED bit
after copying into the write buffer.